+++ src/apps/shell/rust_parallel.rs
@@ -37,8 +37,7 @@ impl App for RustParallel {
 
     fn download(&self) -> Result<AppAssets> {
         let release = self.client.latest_release(Self::OWNER, Self::REPO)?;
-        let name =
-            release.find_asset(|a| a == "rust-parallel-x86_64-unknown-linux-gnu.tar.gz")?;
+        let name = release.find_asset(|a| a == "rust-parallel-x86_64-unknown-linux-gnu.tar.gz")?;
         let asset = self.client.download_asset(Self::OWNER, Self::REPO, &name)?;
         let extractor = ArchiveExtractor::new(&name, asset.data);
         let binary_data = extractor.extract_by_filename(Self::EXE_NAME)?;